I organized the structure of files and screen and components
