--- tic-tac-toe/src/app/components/App/test/index.test.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-import App from '../index';
-
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
-  ReactDOM.unmountComponentAtNode(div);
-});
