+++ tic-tac-toe/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Game from './app/components/game';
+import Game from '../src/app/screens/Game';
 import './scss/index.scss';
 
 ReactDOM.render(<Game />, document.getElementById('root'));
